refactor: move screen to /CreateWallet
