+++ mobile-app/app/screens/WalletNavigator/screens/CreateWallet/OnboardingNetworkSelectScreenV2.test.tsx
